fix loi hien thi icon va hien thi san pham theo danh muc
